requirements/local.txt
-r base.txt
django-mockups
django-extensions
ptpython